cleaned conversion function

+++ main/boards/drivers/TPS53647.h
@@ -29,7 +29,7 @@
 int TPS53647_init(int num_phases, int imax, float ifault);
 int TPS53647_get_frequency(void);
 void TPS53647_set_frequency(int);
-int TPS53647_get_temperature(void);
+float TPS53647_get_temperature(void);
 float TPS53647_get_vin(void);
 float TPS53647_get_iout(void);
 float TPS53647_get_iin(void);
